ImpactAnalyzer/requirements.txt
paramiko
oracledb
websocket-client
requests
